# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bore Tunnel Management API
  description: |
    REST API for managing TCP tunnels through bore-server infrastructure.
    Provides user authentication, instance management, and real-time status monitoring.
    
    ## API Versioning
    
    This API uses URL-based versioning with the `/api/v1/` prefix. All endpoints are versioned
    to ensure backward compatibility and smooth migration paths.
    
    **Current Version:** v1 (Active)
    
    **Versioning Strategy:**
    - New endpoints and optional fields can be added without breaking changes
    - Breaking changes will result in a new major version (e.g., v2)
    - Previous major versions are supported for a minimum of 12 months after deprecation
    - Deprecated versions include `X-API-Deprecation` headers with sunset dates
    
    **Legacy Support:**
    - Unversioned endpoints (e.g., `/api/auth/*`) automatically redirect to `/api/v1/*` using HTTP 308
    - Clients should explicitly use versioned endpoints in production
    
    For detailed versioning documentation, see the [Development Guide](../../DEVELOPMENT.md#api-versioning-strategy).
  version: 1.0.0
  contact:
    name: Bore Support
    email: support@bore.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.bore.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
  schemas:
...
    Instance:
      type: object
      properties:
        id:
          type: string
          example: inst_1234567890_xyz789
        user_id:
          type: string
        name:
          type: string
          example: My Dev Environment
        local_port:
          type: integer
          minimum: 1
          maximum: 65535
          example: 3000
        remote_port:
          type: integer
          nullable: true
        region:
          type: string
          example: us-east
        server_host:
          type: string
        assigned_server:
          type: string
          nullable: true
        status:
          type: string
          enum: [inactive, starting, active, online, idle, degraded, offline]
          example: online
        status_reason:
          type: string
        tunnel_connected:
          type: boolean
        public_url:
          type: string
          nullable: true
        current_tunnel_token:
          type: string
          nullable: true
        tunnel_token_expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TunnelToken:
      type: object
      properties:
        tunnel_token:
          type: string
          example: a1b2c3d4e5f6...
        bore_server_host:
          type: string
          example: bore-server-01.bore.com
        bore_server_port:
          type: integer
          example: 7835
        local_port:
          type: integer
        expires_at:
          type: string
          format: date-time
        server_info:
          type: object
          properties:
            server_id:
              type: string
            utilization:
              type: string
...
paths:
...
  /instances:
    get:
      tags:
        - Instances
      summary: List user instances
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Instance'

    post:
      tags:
        - Instances
      summary: Create new instance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - localPort
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                localPort:
                  type: integer
                  minimum: 1
                  maximum: 65535
                region:
                  type: string
                server_host:
                  type: string
      responses:
        '201':
          description: Instance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '429':
          description: Quota exceeded
        '503':
          description: System at capacity
  /instances/{id}:
    patch:
      tags:
        - Instances
      summary: Rename instance
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Instance updated
        '404':
          description: Instance not found

    delete:
      tags:
        - Instances
      summary: Delete instance
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance deleted
        '404':
          description: Instance not found
  /instances/{id}/heartbeat:
    post:
      tags:
        - Instances
      summary: Send instance heartbeat
      description: Update instance status with health metrics
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                vscode_responsive:
                  type: boolean
                last_activity:
                  type: integer
                cpu_usage:
                  type: number
                  minimum: 0
                  maximum: 100
                memory_usage:
                  type: integer
                has_code_server:
                  type: boolean
      responses:
        '200':
          description: Heartbeat recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: string
                  reason:
                    type: string
  /instances/{id}/connect:
    post:
      tags:
        - Instances
      summary: Get tunnel connection token
      description: Generate token for establishing tunnel connection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tunnel token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TunnelToken'
        '503':
          description: All servers at capacity
  /instances/{id}/disconnect:
    post:
      tags:
        - Instances
      summary: Disconnect tunnel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Disconnected successfully
...
  /instances/{id}/status-history:
    get:
      tags:
        - Instances
      summary: Get instance status history
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status history
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance_id:
                    type: string
                  current_status:
                    type: string
                  status_reason:
                    type: string
                  health_metrics:
                    type: object
                  last_heartbeat:
                    type: integer
                  heartbeat_age_ms:
                    type: integer
                  status_history:
                    type: array
                    items:
                      type: object
                  uptime_data:
                    type: object

  # Monitoring
...